# src/main/resources/OpenAPI.yml
openapi: 3.0.3
info:
  title: Sevyh Chat Media Service
  version: 1.0.0
servers:
  - url: /api/v1/chatmedia
paths:
  /:
    post:
      summary: Create a new message with media
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                senderId:
                  type: string
                  format: uuid
                receiverId:
                  type: string
                  format: uuid
                textContent:
                  type: string
                timestamp:
                  type: string
                  format: date-time
                messageType:
                  $ref: '#/components/schemas/MessageType'
                media:
                  type: string
                  format: binary
                createdAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Chat message created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMessage'
        '400':
          description: Bad request
  /{messageId}:
    get:
      summary: Get a chat message by ID
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chat message retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMessage'
        '404':
          description: Message not found
  /media/{mediaId}:
    get:
      summary: Get media metadata by ID
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Media metadata retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMediaMetadata'
        '404':
          description: Media metadata not found
components:
  schemas:
    MessageType:
      type: string
      enum:
        - TEXT
        - IMAGE
        - VIDEO
    Message:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        textContent:
          type: string
        senderId:
          type: string
          format: uuid
        receiverId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        messageType:
          $ref: '#/components/schemas/MessageType'
    MediaMetadata:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        mediaType:
          type: string
        size:
          type: integer
          format: int64
        fileLocation:
          type: string
    ApiResponseMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Message'
    ApiResponseMediaMetadata:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/MediaMetadata'